fix: There is identical code in two places
