+++ src/routes/Session.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 
 import SessionController from '../controllers/Session';
 import { asyncController } from '../plugins/utility';
-import SessionValidator from '../validator/Session';
+import SessionValidator from '../validator/util';
 
 const router = express.Router();
 const { post } = SessionController;
