+++ src/validator/util.ts
@@ -1,11 +1,18 @@
+import { NextFunction, Request, Response } from 'express';
 import * as createError from 'http-errors';
 import * as Joi from 'joi';
 import * as _ from 'lodash';
-
+import { schemasValidationSession } from './Session';
+import { schemasValidationUser } from './User';
 interface IValidations {
   schema?: Joi.ObjectSchema;
   data: object;
 }
+
+const schemasValidation = {
+  users: schemasValidationUser,
+  sessions: schemasValidationSession
+};
 
 const validation = (options: IValidations): void => {
   const { schema, data } = options;
@@ -21,18 +28,29 @@ const validation = (options: IValidations): void => {
 
   return;
 };
-export const validator = async (
-  schemasValidation: Joi.ObjectSchema,
+const validator = async (
+  schemasValidations: Joi.ObjectSchema,
   body: object,
   method: string
 ) => {
-  //   const validations = {
-  //     POST: () => validation({ schema: schemasValidation.post, data: body }),
-  //     PUT: () => validation({ schema: schemasValidation.put, data: body })
-  //   };
   if (method === 'POST' || method === 'PUT') {
-    await validation({ schema: schemasValidation, data: body });
+    await validation({ schema: schemasValidations, data: body });
   }
 
   return;
+};
+
+export default async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    const method = req.method;
+    const body = req.body;
+    const url = _.trim(req.baseUrl, '/');
+    const schemas = schemasValidation[url];
+
+    await validator(schemas[method], body, method);
+
+    next();
+  } catch (error) {
+    throw error;
+  }
 };
